Añadimos Salida.txt 's

+++ MPI/main.c
@@ -56,8 +56,8 @@ void CopiarDatos3(Celda** estado_viejo, Celda** estado_actual){
         memcpy((*estado_viejo), (*estado_actual), sizeof(Celda)*n*n);
 }*/
 
-float generador_Uniforme(int random, int a, int b){
-    float resultado=((float)(random %(b-a+1) + a)/100.0f);
+double generador_Uniforme(int random, int a, int b){
+    double resultado=((double)(random %(b-a+1) + a)/100.0f);
     return resultado;
 }
 
@@ -81,7 +81,7 @@ void init(Celda** estadoActual , int inicio, int final){
     Celda Celda_auxiliar;
     for(int i=inicio; i<final ; i++) {
         for (int j = 0; j < N; j++) {
-            float prob = generador_Uniforme(rand(),0,100);
+            double prob = generador_Uniforme(rand(),0,100);
             if(prob<=0.05){
                 Celda_auxiliar.estado=ROJO;
                 Celda_auxiliar.tiempo_contagio=(generadorUniformeENTEROS(rand(),0,7));
@@ -164,8 +164,8 @@ MPI_Datatype generarTipo() {
 }
 
 
-float susceptibilidad(int edad,int heridas_A){
-    float suscep=0;
+double susceptibilidad(int edad,int heridas_A){
+    double suscep=0;
     if(edad<=156){
         suscep=0.35;
     }else{
@@ -182,7 +182,7 @@ float susceptibilidad(int edad,int heridas_A){
     return suscep;
 }
 
-float procesarContagio(float Porc_vecinosEnf ,float susceptibilidad){
+double procesarContagio(double Porc_vecinosEnf ,double susceptibilidad){
     return ((Porc_vecinosEnf + susceptibilidad)* 0.60) + 0.07;
 }
 
@@ -198,7 +198,7 @@ Celda procesarCelda(Celda celda, int vecinosEnfermos){
 
     switch (celda.estado) {
         case ROJO:{
-            int probabilidad= generador_Uniforme(rand(),0,100);
+            double probabilidad= generador_Uniforme(rand(),0,100);
             if((celda.tiempo_contagio>4)&&(probabilidad<0.85)){
                 nuevaCelda.estado=AZUL;
             }
@@ -207,7 +207,7 @@ Celda procesarCelda(Celda celda, int vecinosEnfermos){
         }
         case AZUL:{
             if(celda.tiempo_contagio>7){
-                int probabilidad= generador_Uniforme(rand(),0,100);
+                double probabilidad= generador_Uniforme(rand(),0,100);
                 if(celda.edad<156){
                     if(probabilidad<=0.03){
                         nuevaCelda.estado=BLANCO;
@@ -263,7 +263,7 @@ Celda procesarCelda(Celda celda, int vecinosEnfermos){
             break;
         }
         case VERDE:{
-            float probabilidad=procesarContagio(((float)vecinosEnfermos/12),susceptibilidad(celda.edad,celda.herida_abierta));
+            double probabilidad=procesarContagio(((double)vecinosEnfermos/12),susceptibilidad(celda.edad,celda.herida_abierta));
             if(generador_Uniforme(rand(),0,100)<probabilidad){
                 nuevaCelda.estado=NARANJA;
                 nuevaCelda.tiempo_contagio=0;
