+++ OpenMP/main.c
@@ -34,8 +34,8 @@ Celda ** Crear_Matriz(){
 }
 
 
-float generador_Uniforme(int random, int a, int b){
-    float resultado=((float)(random %(b-a+1) + a)/100.0f);
+double generador_Uniforme(int random, int a, int b){
+    double resultado=((double)(random %(b-a+1) + a)/100.0f);
     return resultado;
 }
 
@@ -57,7 +57,7 @@ void init(Celda** estadoActual){
         for(i=0; i<n ; i++) {
             #pragma omp parallel for shared(estadoActual) private(j,Celda_auxiliar) num_threads(3)
             for (j = 0; j < n; j++) {
-                float prob = generador_Uniforme(rand(),0,100);
+                double prob = generador_Uniforme(rand(),0,100);
                 if(prob<=0.05){
                     Celda_auxiliar.estado=ROJO;
                     Celda_auxiliar.tiempo_contagio=(generadorUniformeENTEROS(rand(),0,7));
@@ -111,8 +111,8 @@ void init(Celda** estadoActual){
         }
     }
 
-float susceptibilidad(int edad,int heridas_A){
-    float suscep=0;
+double susceptibilidad(int edad,int heridas_A){
+    double suscep=0;
     if(edad<=156){
         suscep=0.35;
     }else{
@@ -129,7 +129,7 @@ float susceptibilidad(int edad,int heridas_A){
     return suscep;
 }
 
-float procesarContagio(float Porc_vecinosEnf ,float susceptibilidad){
+double procesarContagio(double Porc_vecinosEnf ,double susceptibilidad){
     return ((Porc_vecinosEnf + susceptibilidad)* 0.60) + 0.07;
 }
 
@@ -145,7 +145,7 @@ Celda procesarCelda(Celda celda, int vecinosEnfermos){
 
     switch (celda.estado) {
         case ROJO:{
-            int probabilidad= generador_Uniforme(rand(),0,100);
+            double probabilidad= generador_Uniforme(rand(),0,100);
             if((celda.tiempo_contagio>4)&&(probabilidad<0.85)){
                 nuevaCelda.estado=AZUL;
             }
@@ -154,7 +154,7 @@ Celda procesarCelda(Celda celda, int vecinosEnfermos){
         }
         case AZUL:{
             if(celda.tiempo_contagio>7){
-                int probabilidad= generador_Uniforme(rand(),0,100);
+                double probabilidad= generador_Uniforme(rand(),0,100);
                 if(celda.edad<156){
                     if(probabilidad<=0.03){
                         nuevaCelda.estado=BLANCO;
@@ -210,7 +210,7 @@ Celda procesarCelda(Celda celda, int vecinosEnfermos){
             break;
         }
         case VERDE:{
-            float probabilidad=procesarContagio(((float)vecinosEnfermos/12),susceptibilidad(celda.edad,celda.herida_abierta));
+            double probabilidad=procesarContagio(((double)vecinosEnfermos/12),susceptibilidad(celda.edad,celda.herida_abierta));
             if(generador_Uniforme(rand(),0,100)<probabilidad){
                 nuevaCelda.estado=NARANJA;
                 nuevaCelda.tiempo_contagio=0;
@@ -378,9 +378,9 @@ int main() {
     double tiempo_total=0;
     double promedio;
    /* struct timeval tiempoInicial, tiempoFinal;
-    float tiempoTotal;
-    float sumaTiempo=0;
-    float promedio = 0;*/
+    double tiempoTotal;
+    double sumaTiempo=0;
+    double promedio = 0;*/
     Celda ** Estado_actual = Crear_Matriz();
     Celda ** Estado_siguiente = Crear_Matriz();
     Celda ** Aux;
